Move VerificationResult and CommandLineParseState to separate files
